+++ react/Drawer.tsx
@@ -1,5 +1,4 @@
 import React, {
-  ReactElement,
   Suspense,
   useReducer,
   MouseEventHandler,
@@ -7,15 +6,20 @@ import React, {
   useState,
 } from 'react'
 import { defineMessages } from 'react-intl'
-import { useCssHandles } from 'vtex.css-handles'
 import { IconMenu } from 'vtex.store-icons'
+import { useCssHandles } from 'vtex.css-handles'
 import { useChildBlock, ExtensionPoint } from 'vtex.render-runtime'
+import {
+  MaybeResponsiveValue,
+  useResponsiveValue,
+} from 'vtex.responsive-values'
 
-import Overlay from './Overlay'
 import Portal from './Portal'
+import Overlay from './Overlay'
 import useLockScroll from './modules/useLockScroll'
 import DrawerCloseButton from './DrawerCloseButton'
 import { DrawerContextProvider } from './DrawerContext'
+import { isElementInsideLink } from './modules/isElementInsideLink'
 
 const Swipable = React.lazy(() => import('./Swipable'))
 
@@ -31,6 +35,27 @@ interface MenuAction {
 const initialMenuState: MenuState = {
   isOpen: false,
   hasBeenOpened: false,
+}
+
+type Position = 'left' | 'right' | 'up' | 'down'
+type SlideDirection = 'vertical' | 'horizontal' | 'rightToLeft' | 'leftToRight'
+type Height = '100%' | 'auto' | 'fullscreen'
+type Width = '100%' | 'auto'
+type BackdropMode = 'visible' | 'none'
+
+interface Props {
+  actionIconId: string
+  dismissIconId: string
+  position: Position
+  width?: Width
+  height: Height
+  slideDirection?: SlideDirection
+  isFullWidth: boolean
+  maxWidth?: number | string
+  children: React.ReactNode
+  customIcon: React.ReactElement
+  header: React.ReactElement
+  backdropMode?: MaybeResponsiveValue<BackdropMode>
 }
 
 function menuReducer(state: MenuState, action: MenuAction) {
@@ -79,57 +104,21 @@ const CSS_HANDLES = [
   'drawerContent',
   'childrenContainer',
   'closeIconContainer',
-]
+] as const
 
-// This is a totally valid use case for any, eslint.
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-function isHTMLElement(x: any): x is HTMLElement {
-  return 'tagName' in x
-}
-
-function isLink(element: HTMLElement): element is HTMLAnchorElement {
-  return String(element.tagName).toUpperCase() === 'A'
-}
-
-const isElementInsideLink = (
-  node: HTMLElement | null,
-  limit?: HTMLElement
-): boolean => {
-  if (!node || !isHTMLElement(node)) {
-    return false
-  }
-
-  if (isLink(node)) {
-    return true
-  }
-
-  const { parentNode } = node
-
-  if (
-    !parentNode ||
-    !isHTMLElement(parentNode) ||
-    parentNode.tagName.toUpperCase() === 'BODY' ||
-    (limit && parentNode === limit)
-  ) {
-    return false
-  }
-
-  return isElementInsideLink(parentNode, limit)
-}
-
-const Drawer: StorefrontComponent<DrawerSchema & {
-  customIcon?: ReactElement
-  header?: ReactElement
-}> = ({
-  width,
-  customIcon,
-  maxWidth = 450,
-  isFullWidth,
-  slideDirection = 'horizontal',
-  header,
-  children,
-}) => {
+function Drawer(props: Props) {
+  const {
+    width,
+    header,
+    children,
+    customIcon,
+    isFullWidth,
+    maxWidth = 450,
+    slideDirection = 'horizontal',
+    backdropMode: backdropModeProp = 'visible',
+  } = props
   const handles = useCssHandles(CSS_HANDLES)
+  const backdropMode = useResponsiveValue(backdropModeProp)
   const hasTriggerBlock = Boolean(useChildBlock({ id: 'drawer-trigger' }))
   const hasHeaderBlock = Boolean(useChildBlock({ id: 'drawer-header' }))
   const { state: menuState, openMenu, closeMenu } = useMenuState()
@@ -162,12 +151,15 @@ const Drawer: StorefrontComponent<DrawerSchema & {
     [isMenuOpen, openMenu, closeMenu]
   )
 
+  const overlayVisible = backdropMode === 'visible' && isMenuOpen
+
   return (
     <DrawerContextProvider value={contextValue}>
       <div
-        className={`pa4 pointer ${handles.openIconContainer}`}
         onClick={openMenu}
-        aria-hidden
+        role="presentation"
+        aria-hidden={isMenuOpen ? 'false' : 'true'}
+        className={`pa4 pointer ${handles.openIconContainer}`}
       >
         {hasTriggerBlock ? (
           <ExtensionPoint id="drawer-trigger" />
@@ -176,7 +168,7 @@ const Drawer: StorefrontComponent<DrawerSchema & {
         )}
       </div>
       <Portal>
-        <Overlay visible={isMenuOpen} onClick={closeMenu} />
+        <Overlay visible={overlayVisible} onClick={closeMenu} />
         <Suspense fallback={<React.Fragment />}>
           <Swipable
             {...{
@@ -194,7 +186,7 @@ const Drawer: StorefrontComponent<DrawerSchema & {
               direction === 'right' ? 'right-0' : 'left-0'
             } fixed top-0 bottom-0 bg-base z-999 flex flex-column`}
             style={{
-              width: width || (isFullWidth ? '100%' : '85%'),
+              width: width ?? (isFullWidth ? '100%' : '85%'),
               maxWidth,
               minWidth: 280,
               pointerEvents: isMenuOpen ? 'auto' : 'none',